Move internal plugins to top-level
Thought about moving them into standalone modules but there's no real
benefit to that beyond my own module count vanity
