+++ lib/runner.js
@@ -5,11 +5,11 @@ var resultsStore = require('./report/results-store')
 var errorStore = require('./report/error-store')
 var pluginsStore = require('./plugins/store')
 
-var donePlugin = require('./plugins/internal/done')
-var uncaughtExceptionPlugin = require('./plugins/internal/uncaught-exception')
-var timeoutPlugin = require('./plugins/internal/timeout')
-var resultsPlugin = require('./plugins/internal/results')
-var tap13Plugin = require('./plugins/internal/tap13')
+var donePlugin = require('../plugins/done')
+var uncaughtExceptionPlugin = require('../plugins/uncaught-exception')
+var timeoutPlugin = require('../plugins/timeout')
+var resultsPlugin = require('../plugins/results')
+var tap13Plugin = require('../plugins/tap13')
 
 module.exports = function (action, log, timeout, cb) {
   resultsStore.initialize()
